+++ mmg/cli.py
@@ -92,20 +92,22 @@ def mmgcli(
     verbose: int,
 ):
     """
-    FILE_NAMES: Base file names to convert. `*.base.md` or `*.base.ipynb` are available.
+    FILE_NAMES: Base file names to convert.
+
+    Supported extensions are:
+        .md, .markdown, .mkd (Markdown)
+        .mdx (MDX - Markdown + JSX) [experimental]
+        .rmd (R Markdown) [experimental]
+        .mmd (MultiMarkdown) [experimental]
+        .qmd (Quarto) [experimental]
+        .ipynb (Jupyter Notebook)
 
     Here are some examples:
-
         mmg *.base.md
-
         mmg *.base.ipynb
-
         mmg *.base.md *.base.ipynb -o pdf --css github-dark
-
         mmg --recursive
-
         mmg --recursive --validation-only
-
         mmg --batch mmg.yml
     """
     if file_names or recursive: